+++ tests/testthat/reporters/context.R
@@ -1,3 +1,4 @@
+testthat:::local_edition(2)
 context("my context")
 
 test_that("a test", {